워치 공유 모델 분리
